# Conflicts:
#	sumaris-core/src/main/java/net/sumaris/core/config/CacheConfiguration.java
#	sumaris-core/src/test/java/net/sumaris/core/dao/referential/TaxonNameRepositoryReadTest.java
#	sumaris-server/src/main/java/net/sumaris/server/http/graphql/data/DataGraphQLService.java
#	sumaris-server/src/main/java/net/sumaris/server/http/graphql/referential/ReferentialGraphQLService.java

+++ sumaris-core-extraction/src/main/java/net/sumaris/core/extraction/dao/technical/ExtractionBaseDaoImpl.java
@@ -116,9 +116,9 @@ public abstract class ExtractionBaseDaoImpl extends HibernateDaoSupport {
         return resultStream.map(rowMapper).collect(Collectors.toList());
     }
 
-    protected <R> List<R> query(String query, Function<Object[], R> rowMapper, int offset, int size) {
+    protected <R> List<R> query(String query, Function<Object[], R> rowMapper, long offset, int size) {
         Query nativeQuery = createNativeQuery(query)
-                .setFirstResult(offset)
+                .setFirstResult((int)offset)
                 .setMaxResults(size);
         Stream<Object[]> resultStream = (Stream<Object[]>) nativeQuery.getResultStream();
         return resultStream.map(rowMapper).collect(Collectors.toList());
